% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_functions.R
\name{discrimination_law}
\alias{discrimination_law}
\title{Discrimination Law Function}
\usage{
discrimination_law(country)
}
\arguments{
\item{country}{A string representing country listed in the dataset column "Country".}
}
\value{
A short word or phrase describing whether laws prevent forms of discrimination based on gender identity or sexuality.
}
\description{
Display whether laws prevent LGBTQ+ discrimination in a given country.
}
\note{
Using this package requires the dataset "PTNS_countries_LGBTQ.csv" which can be found in the package on GitHub.
}
\examples{
# Check whether laws exist to prevent discrimination in Mexico.
discrimination_law("Canada")

}
