% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_functions.R
\name{gender_change_law}
\alias{gender_change_law}
\title{Gender Change Law Function}
\usage{
gender_change_law(country)
}
\arguments{
\item{country}{A string representing country listed in the dataset column "Country".}
}
\value{
A short word or phrase describing if/how legal gender change is possible.
}
\description{
Display under what conditions legal gender change is possible in a given country.
}
\note{
Using this package requires the dataset "PTNS_countries_LGBTQ.csv" which can be found in the package on GitHub.
}
\examples{
# Check whether gender change is legal in Canada.
gender_change_law("Canada")

}
