% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_functions.R
\name{open_app}
\alias{open_app}
\title{Open App Function}
\usage{
open_app()
}
\description{
Runs the LGBTQsafety app in R shiny, where the previous functions are utilized with a GUI.
}
\note{
This function requires the "shiny", "bslib", "dplyr", and "LGBTQsafety" packages to be loaded.
}
\examples{
# Run the app in R shiny
open_app()

}
