% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_functions.R
\name{read_and_calculate_data}
\alias{read_and_calculate_data}
\title{Read and Calculate Data Function}
\usage{
read_and_calculate_data(start)
}
\arguments{
\item{start}{A string where users should enter the word "yes".}
}
\value{
If users enter "yes", this returns the dataset and saves it. If not, users are prompted to enter "yes".
}
\description{
Reads the data set required for the following functions, converts string values in the columns same.sex.sexual.activity, same.sex.marriage, gender.ID.expression and anti.discrimination.laws.sexual.orientation for calculating safety ratings, and downloads the dataset to the global environment.
}
\note{
Using this package requires the dataset "PTNS_countries_LGBTQ.csv", and also requires the dplyr package.
}
\examples{
# Download the dataset
read_and_calculate_data("yes")

}
