% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_functions.R
\name{read_data}
\alias{read_data}
\title{Read Data Function}
\usage{
read_data(start)
}
\arguments{
\item{start}{A string where users should enter the word "yes".}
}
\value{
If users enter "yes", this returns the dataset and saves it. If not, users are prompted to enter "yes".
}
\description{
Reads the dataset required for the following functions in the package and downloads it to the global environment.
}
\examples{
# Download the dataset
read_data("yes")

}
