% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_functions.R
\name{safety_rating}
\alias{safety_rating}
\title{Safety Rating Function}
\usage{
safety_rating(country)
}
\arguments{
\item{country}{A string representing country listed in the dataset column "Country".}
}
\value{
A number between 1 and 5 which represents a safety rating.
}
\description{
Displays safety rating of a given country on a scale from 1 to 5. This is calculated using laws regarding legal status of homosexuality, same-sex marriage, gender change, and discrimination.
}
\note{
Using this package requires the dataset "PTNS_countries_LGBTQ.csv" which can be found in the package on GitHub.
}
\examples{
# Check the overall safety rating for LGBTQ+ individuals in China.
safety_rating("China")

}
