% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_functions.R
\name{sexuality_marriage_law}
\alias{sexuality_marriage_law}
\title{Sexuality Law Function}
\usage{
sexuality_marriage_law(country)
}
\arguments{
\item{country}{A string representing a country listed in the dataset column "Country".}
}
\value{
A short word or phrase describing whether gay marriage is legal or illegal.
}
\description{
Display whether same-sex marriage is legal in a given country
}
\note{
Using this package requires the dataset "PTNS_countries_LGBTQ.csv" which can be found in the package on GitHub.
}
\examples{
# Check whether same sex marriage is legal in Uganda
sexuality_marriage_law("Uganda")

}
